+++ src/queries/events/deleteEvents.js
@@ -24,7 +24,15 @@ async function deleteDuplicatesForCourses() {
 	return db.query(query);
 }
 
+async function deleteEventsForScope(scopeId) {
+	const query = `DELETE FROM events WHERE scope_id = $1 RETURNING ${allColumns}`;
+	const params = [scopeId];
+
+	return db.query(query, params);
+}
+
 module.exports = {
 	deleteEvents,
+	deleteEventsForScope,
 	deleteDuplicatesForCourses,
 };